// PCDACO-Owner/constants/models/report.model.ts
export interface ReportApprovePayload {
  isApproved: boolean;
  note: string;
}

export interface ReportApproveResponse {
  id: string;
  title: string;
  description: string;
  status: string;
  resolutionComments: string;
  resolvedAt: Date;
  isCompensationPaid: boolean;
  compensationAmount: number;
  compensationReason: string;
}

export interface ReportCompensationPayload {
  userId: string;
  compensationReason: string;
  compensationAmount: number;
}

export interface ReportCompensationResponse {
  reportId: string;
}

export interface ReportPayload {
  bookingId: string;
  title: string;
  description: string;
  reportType: number;
}

export interface ReportResponse {
  id: string;
}

export interface ReportListResponse {
  id: string;
  bookingId: string;
  reporterId: string;
  reporterName: string;
  title: string;
  description: string;
  reportType: number;
  status: number;
  resolvedAt: null;
  resolvedById: null;
  resolutionComments: null;
  imageReports: ImageReport[];
}

export interface ImageReport {
  id: string;
  url: string;
}

export interface ReportParams extends RootRequest {
  status: number;
  type: number;
}

export interface ReportDetailResponse {
  id: string;
  reporterId: string;
  reporterName: string;
  reporterRole: string;
  title: string;
  description: string;
  reportType: number;
  status: number;
  resolvedAt: null;
  resolvedById: null;
  resolutionComments: null;
  imageUrls: string[];
  bookingDetail: BookingDetail;
  carDetail: CarDetail;
  compensationDetail: CompensationDetail;
  inspectionScheduleDetail: InspectionScheduleDetail;
}

export interface BookingDetail {
  id: string;
  driverId: string;
  driverName: string;
  driverAvatar: string;
  driverPhone: string;
  ownerId: string;
  ownerName: string;
  ownerAvatar: string;
  ownerPhone: string;
  startTime: Date;
  endTime: Date;
  totalAmount: number;
  basePrice: number;
}

export interface CarDetail {
  id: string;
  licensePlate: string;
  modelName: string;
  manufacturerName: string;
  color: string;
  imageUrl: string[];
}

export interface CompensationDetail {
  userId: string;
  userName: string;
  userAvatar: string;
  compensationReason: string;
  compensationAmount: number;
  isPaid: boolean;
  imageUrl: string;
  paidAt: Date;
}

export interface InspectionScheduleDetail {
  id: string;
  technicianId: string;
  technicianName: string;
  technicianAvatar: string;
  status: number;
  inspectionAddress: string;
  inspectionDate: Date;
  note: string;
  photoUrls: string[];
}

export interface ReportCompensationProofPayload {
  images: File;
}

export interface ReportReviewResponse {
  reportId: string;
  status: string;
}

export interface ReportCompensationProofResponse {
  reportId: string;
  imageUrl: string;
}

export interface ReportImagePayload {
  files: File[];
}

// Generated by https://quicktype.io

export interface ReportImageReponse {
  images: Image[];
}

export interface Image {
  id: string;
  url: string;
}